Web_Demo/requirements.txt
opencv-python 
numpy 
scikit-image 
torch 
torchvision
streamlit
Pillow
pickle-mixin
scikit-learn

